Add example news links to HP and clean up
